requirements.txt
Flask
Pillow
requests
python-dotenv
gunicorn 